Prepare for release
This adds a changelog, updates package.json with more up to date
information (also includes a reference to the actual index file).

It update the tsconfig file to produce a dist folder that has only what
is needed (not shipping the specs or anything).

I'm also adding a release script and an .npmignore to make sure we only
release what we need.