Only allow Joi null/undefined when passed a null/undefined value
Previously it was generating the Joi schema `Joi.allow(null)` for null
values, this is an alias for `Joi.any.allow(null)`. Allow is for
specifying values that are accepted in addition to the already defined
filter, and since the default filter is `any` it means that these
filters will accept any value.

`Joi.valid` seems like what it should be instead as `valid` restricts
the schema to only the values specified.