Move the recursion handling to the transpiler
Keeping track of recuring and self referencing types was done in the
modules up until now. This meant that the transpiler was not really
aware of the types that were already resolved. Because of that it was
not possible to introduce something like eager references (which would
reduce the size of the output schemas).

This commit moves all the logic around visited types and references to
the transpiler.