Add support for types that have the same name
It turns out that if a file defines something like:
```
import { A, B } from './sharedTypes'
type Output = A | B
```
and another unrelated file (i.e. a file that does not import the first
one) defines
```
import { A, B } from './sharedTypes'
type Output = Array<A | B>
```

If the files are both on the compiler instance, typescript will resolve
the second `Output` as `type Output = Array<Output>` where the first
`Output` has the same name `checker.typeToString(type)` as the second
but a different type id.

It kind of makes sens but it means we have to handle that and can't just
use the typeString to keep track of references.