Keep track of names used in a file
It is an assumption that each file processed will be outputed to a file
so we have to ensure that accross the whole file names are unique.
Before this change the following file structure would yield nonsensical
results:

./a.ts
```ts
type Input = { id: number }
type Output = {} | null
```

./b.ts
```ts
type Input = {} | null
type Output = { names: string[] }
```

If both those files are added to the compiler instance, then, when
resolving a.ts with eager refs we get the following output:
```ts
const Input = Joi.alternatives([Joi.valid(null), Joi.object({})])
const input = Joi.lazy(() => Input)
const output = Joi.lazy(() => Input)
```

This is because the resolver first resolved `input`, the result is the
following:
```ts
const Input = Joi.object({ id: Joi.number() })
const input = Joi.lazy(() => Input)
```
Then when resolving `output` it forgets about the fact that the name
`Input` is used.

Because `./b.ts` defines `type Input = {} | null` then internally to the
TS checker, when the name of `type Output = {} | null` vaguely resolves
to `type Ouput = Input` (oversimplified in some way). That input
reference is for the type from `./b.ts` but because we're not aware that
`Input` is used we end up overriding the first input leading to the
nonsensical result.

An alternative would have been to be less eager but a similar issue
could have cropped up.

It would be best to keep all encountered types in memory but for now
this should be enough